requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.4
Django==4.2.6
django-cors-headers==4.3.1
django-filter==23.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycparser==2.21
pyscopg2==66.0.2
pytz==2023.3.post1
sqlparse==0.4.4
